refactor(193): update architecture for StudentTrackCard
