+++ src/features/student/components/cards/index.ts
@@ -1,2 +1,1 @@
 export * from './StudentSkillCard'
-export * from './StudentTrackCard'
