+++ src/features/student/components/widgets/StudentTracksWidget/components/StudentTrackCard/StudentTrackCard.vue
